Initial typescript conversion of rtf.js
