+++ package.json
@@ -1,33 +1,46 @@
 {
-  "name": "rtf.js",
-  "version": "1.0.0",
-  "description": "Render RTF documents in HTML. This also includes rendering WMF and EMF images which are often contained in RTF documents.",
-  "main": "rtf.js",
-  "scripts": {
-    "lint": "eclint check *.js test/*.js",
-    "cleanup": "eclint fix *.js test/*.js",
-    "generate-testcase": "node test/generate-testcase.js",
-    "test": "mocha test/test-rtfjs.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/tbluemel/rtf.js.git"
-  },
-  "keywords": [
-    "RTF"
-  ],
-  "author": "Thomas Bluemel",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/tbluemel/rtf.js/issues"
-  },
-  "homepage": "https://github.com/tbluemel/rtf.js#readme",
-  "devDependencies": {
-    "chai": "^4.1.2",
-    "chai-html": "^1.1.5",
-    "eclint": "^2.6.0",
-    "jsdom": "^11.6.2",
-    "mocha": "^5.0.1",
-    "pretty": "^2.0.0"
-  }
+    "name": "rtf.js",
+    "version": "1.0.0",
+    "description": "Render RTF documents in HTML. This also includes rendering WMF and EMF images which are often contained in RTF documents.",
+    "main": "rtf.js",
+    "scripts": {
+        "lint": "eclint check src/* src/**/* test/*.js",
+        "fix": "eclint fix src/* src/**/* test/*.js",
+        "clean": "rimraf dist",
+        "compile": "tsc",
+        "bundle": "rollup --config",
+        "build": "npm run clean && npm run compile && npm run bundle",
+        "generate-testcase": "node test/generate-testcase.js",
+        "test": "mocha test/test-rtfjs.js"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/tbluemel/rtf.js.git"
+    },
+    "keywords": [
+        "RTF"
+    ],
+    "author": "Thomas Bluemel",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/tbluemel/rtf.js/issues"
+    },
+    "homepage": "https://github.com/tbluemel/rtf.js#readme",
+    "devDependencies": {
+        "@types/jquery": "^3.3.0",
+        "chai": "^4.1.2",
+        "chai-html": "^1.1.5",
+        "eclint": "^2.6.0",
+        "jsdom": "^11.6.2",
+        "mocha": "^5.0.1",
+        "pretty": "^2.0.0"
+        "rimraf": "^2.6.2",
+        "rollup": "^0.56.2",
+        "rollup-plugin-commonjs": "^8.3.0",
+        "rollup-plugin-node-resolve": "^3.0.3",
+        "typescript": "2.7.2"
+    },
+    "dependencies": {
+        "codepage": "^1.12.0"
+    }
 }
