+++ src/wmf.js
