+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "compilerOptions": {
+        "outDir": "./dist",
+        "allowJs": true,
+        "target": "es5",
+        "moduleResolution": "node",
+        "module": "es2015"
+    },
+    "include": [
+        "./src/**/*"
+    ]
+}
